Move stac views to own file
