+++ pygeoapi/api/__init__.py
@@ -3093,177 +3093,6 @@ class API:
             return headers, HTTPStatus.BAD_REQUEST, to_json(
                 data, self.pretty_print)
 
-    @gzip
-    @pre_process
-    @jsonldify
-    def get_stac_root(
-            self, request: Union[APIRequest, Any]) -> Tuple[dict, int, str]:
-        """
-        Provide STAC root page
-
-        :param request: APIRequest instance with query params
-
-        :returns: tuple of headers, status code, content
-        """
-
-        if not request.is_valid():
-            return self.get_format_exception(request)
-        headers = request.get_response_headers(**self.api_headers)
-
-        id_ = 'pygeoapi-stac'
-        stac_version = '1.0.0-rc.2'
-        stac_url = f'{self.base_url}/stac'
-
-        content = {
-            'id': id_,
-            'type': 'Catalog',
-            'stac_version': stac_version,
-            'title': l10n.translate(
-                self.config['metadata']['identification']['title'],
-                request.locale),
-            'description': l10n.translate(
-                self.config['metadata']['identification']['description'],
-                request.locale),
-            'links': []
-        }
-
-        stac_collections = filter_dict_by_key_value(self.config['resources'],
-                                                    'type', 'stac-collection')
-
-        for key, value in stac_collections.items():
-            content['links'].append({
-                'rel': 'child',
-                'href': f'{stac_url}/{key}?f={F_JSON}',
-                'type': FORMAT_TYPES[F_JSON]
-            })
-            content['links'].append({
-                'rel': 'child',
-                'href': f'{stac_url}/{key}',
-                'type': FORMAT_TYPES[F_HTML]
-            })
-
-        if request.format == F_HTML:  # render
-            content = render_j2_template(self.tpl_config,
-                                         'stac/collection.html',
-                                         content, request.locale)
-            return headers, HTTPStatus.OK, content
-
-        return headers, HTTPStatus.OK, to_json(content, self.pretty_print)
-
-    @gzip
-    @pre_process
-    @jsonldify
-    def get_stac_path(self, request: Union[APIRequest, Any],
-                      path) -> Tuple[dict, int, str]:
-        """
-        Provide STAC resource path
-
-        :param request: APIRequest instance with query params
-
-        :returns: tuple of headers, status code, content
-        """
-
-        if not request.is_valid():
-            return self.get_format_exception(request)
-        headers = request.get_response_headers(**self.api_headers)
-
-        dataset = None
-        LOGGER.debug(f'Path: {path}')
-        dir_tokens = path.split('/')
-        if dir_tokens:
-            dataset = dir_tokens[0]
-
-        stac_collections = filter_dict_by_key_value(self.config['resources'],
-                                                    'type', 'stac-collection')
-
-        if dataset not in stac_collections:
-            msg = 'Collection not found'
-            return self.get_exception(HTTPStatus.NOT_FOUND, headers,
-                                      request.format, 'NotFound', msg)
-
-        LOGGER.debug('Loading provider')
-        try:
-            p = load_plugin('provider', get_provider_by_type(
-                stac_collections[dataset]['providers'], 'stac'))
-        except ProviderConnectionError as err:
-            LOGGER.error(err)
-            msg = 'connection error (check logs)'
-            return self.get_exception(
-                HTTPStatus.INTERNAL_SERVER_ERROR, headers,
-                request.format, 'NoApplicableCode', msg)
-
-        id_ = f'{dataset}-stac'
-        stac_version = '1.0.0-rc.2'
-
-        content = {
-            'id': id_,
-            'type': 'Catalog',
-            'stac_version': stac_version,
-            'description': l10n.translate(
-                stac_collections[dataset]['description'], request.locale),
-            'links': []
-        }
-        try:
-            stac_data = p.get_data_path(
-                f'{self.base_url}/stac',
-                path,
-                path.replace(dataset, '', 1)
-            )
-        except ProviderNotFoundError as err:
-            LOGGER.error(err)
-            msg = 'resource not found'
-            return self.get_exception(HTTPStatus.NOT_FOUND, headers,
-                                      request.format, 'NotFound', msg)
-        except Exception as err:
-            LOGGER.error(err)
-            msg = 'data query error'
-            return self.get_exception(
-                HTTPStatus.INTERNAL_SERVER_ERROR, headers,
-                request.format, 'NoApplicableCode', msg)
-
-        if isinstance(stac_data, dict):
-            content.update(stac_data)
-            content['links'].extend(stac_collections[dataset]['links'])
-
-            if request.format == F_HTML:  # render
-                content['path'] = path
-                if 'assets' in content:  # item view
-                    if content['type'] == 'Collection':
-                        content = render_j2_template(
-                            self.tpl_config,
-                            'stac/collection_base.html',
-                            content,
-                            request.locale
-                        )
-                    elif content['type'] == 'Feature':
-                        content = render_j2_template(
-                            self.tpl_config,
-                            'stac/item.html',
-                            content,
-                            request.locale
-                        )
-                    else:
-                        msg = f'Unknown STAC type {content.type}'
-                        LOGGER.error(msg)
-                        return self.get_exception(
-                            HTTPStatus.INTERNAL_SERVER_ERROR,
-                            headers,
-                            request.format,
-                            'NoApplicableCode',
-                            msg)
-                else:
-                    content = render_j2_template(self.tpl_config,
-                                                 'stac/catalog.html',
-                                                 content, request.locale)
-
-                return headers, HTTPStatus.OK, content
-
-            return headers, HTTPStatus.OK, to_json(content, self.pretty_print)
-
-        else:  # send back file
-            headers.pop('Content-Type', None)
-            return headers, HTTPStatus.OK, stac_data
-
     def get_exception(self, status, headers, format_, code,
                       description) -> Tuple[dict, int, str]:
         """
